% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setupCovariates}
\alias{setupCovariates}
\title{Set up model covariates}
\usage{
setupCovariates(model, covariates)
}
\arguments{
\item{model}{an instance of \link[OpenMx:MxModel-class]{MxModel}}

\item{covariates}{a character vector naming covariates available in the model data}
}
\value{
The given \link[OpenMx:MxModel-class]{MxModel} with paths
added from covariates to manifest indicators.
}
\description{
In GWAS, including a number of the first principle components as
covariates helps reduce false positives caused by population
stratification. This function adds paths from covariates to
manifest indicators. Covariates are always treated as continuous
variables (not ordinal).
}
\details{
This is not the only way to adjust a model for
covariates. For example, in a single factor model (e.g., \link{buildOneFac}),
it would be more
appropriate to adjust the latent factor instead of the manifest
indicators. However, covariate adjustments to latent variables are only
possible with a maximum likelihood fit function
(\link[OpenMx]{mxFitFunctionML}).  For
\link[OpenMx]{mxFitFunctionWLS}, only manifest indicators can be
adjusted for covariates.
This function always adjusts manifest indicators regardless of the fit function.

You generally do not need to call this function directly because it
is already called by \link{buildOneFac} and similar. This
function is provided for advanced users who wish to write their own
model building functions.
}
\examples{
m1 <- mxModel("test", type="RAM",
             latentVars = "sex", manifestVars = "anxiety")
m1 <- setupCovariates(m1, 'sex')
}
